+++ src/main_loop/main_loop_sdl.c
@@ -202,7 +202,7 @@ static ret_t main_loop_sdl2_dispatch_window_event(main_loop_simple_t* loop, SDL_
         e = event_init(EVT_NATIVE_WINDOW_DESTROY, NULL);
         window_manager_dispatch_native_window_event(l->wm, &e, win);
 
-        main_loop_quit(&(loop->base));
+        tk_quit();
       }
     } break;
 #if SDL_VERSION_ATLEAST(2, 0, 5)